requirements.txt
bluefin-v2-client
python-dotenv
requests